+++ R/wrangling_data.R
@@ -53,21 +53,6 @@ reunite <- function(..., sep = "_") {
   out
 }
 
-#' Moving variables relative to others
-#'
-#' @param data First variable to be used, required.
-#' @param tomove Variable(s) to be moved
-#' @param where String that dictates position in relation to reference variable.
-#' Can be one of: "last", "first", "before", or "after".
-#' @param ref Optional string identifying reference variable
-#' @details Moves variables (columns) of a data frame to positions
-#' relative to other variables in the data frame.
-#' @return The data frame given by 'data' with the variables repositioned
-#' @export
-rearrange <- function(data, tomove, where = "last", ref = NULL) {
-  .Deprecated("dplyr::relocate")
-}
-
 #' Pastes unique string vectors
 #'
 #' For use with dplyr::summarise, for example
@@ -88,34 +73,6 @@ recollect <- function(x, collapse = "_") {
   na_if(paste(unique(na.omit(x)), collapse = collapse), "")
 }
 
-#' Interleaving two vectors by position
-#'
-#' Insert elements in different positions for vectors  
-#' @param vect Main vector
-#' @param pos Positions to be inserted
-#' @param elems Elements to be inserted at those positions.
-#' By default, these are NAs (missing values).
-#' @return A vector the length of the sum of \code{vect}
-#' and \code{pos}.
-#' @examples
-#' interleave(1:5, c(2,4))
-#' @export
-interleave <- function(vect, pos, elems = NA) {
-
-  l <- length(vect)
-  j <- 0
-  for (i in 1:length(pos)) {
-    if (pos[i] == 1)
-      vect <- c(elems[j + 1], vect)
-    else if (pos[i] == length(vect) + 1)
-      vect <- c(vect, elems[j + 1])
-    else
-      vect <- c(vect[1:(pos[i] - 1)], elems[j + 1], vect[(pos[i]):length(vect)])
-    j <- j + 1
-  }
-  return(vect)
-}
-
 #' Get first non-missing
 #'
 #' For use with dplyr::summarise, for example
@@ -126,9 +83,37 @@ interleave <- function(vect, pos, elems = NA) {
 #' but on observations rather than variables.
 #' @source https://stackoverflow.com/questions/40515180/dplyr-how-to-find-the-first-non-missing-string-by-groups
 #' @examples
-#' dplyr::summarise(mtcars, consolidate(mtcars))
-#' consolidate(mtcars$wt)
+#' dplyr::summarise(mtcars, coalesce_rows(mtcars))
+#' coalesce_rows(mtcars$wt)
 #' @export
-consolidate <- function(x) {
+coalesce_rows <- function(x) {
   x[which(!is.na(x))[1]]
+}
+
+#' Fills missing data by lookup
+#'
+#' Fills missing data where known by other observations with the same id/index
+#' @param df a dataframe
+#' @param id a string identifying a column in the dataframe for indexing
+#' @param var a string identifying a column or columns in the dataframe
+#' to be filled
+#' @return a dataframe
+#' @examples
+#' data <- data.frame(ID = c(1,2,3,3,2,1),
+#'                     One = c(1,NA,3,NA,2,NA),
+#'                     Two = c(NA,"B",NA,"C",NA,"A"))
+#' repaint(data, "ID", c("One","Two"))
+#' @export
+repaint <- function(df, id, var) {
+  for (co in var) {
+    for (ea in unique(df[, id])) {
+      if (any(!is.na(df[df[, id] == ea, co])) &
+          any(is.na(df[df[, id] == ea, co]))) {
+        df[df[, id] == ea &
+             is.na(df[, co]), co] <- df[df[, id] == ea &
+                                          !is.na(df[, co]), co][1]
+      }
+    }
+  }
+  df
 }