+++ tests/testthat/test_report_data.R
@@ -1,39 +1,11 @@
-test_that("data_source() returns the correct output format at the package
-          level", {
-    expect_type(data_source(pkg = "qStates", print = FALSE), "list")
-          })
-
-test_that("data_source() returns the correct output format at the database
-          level", {
-  expect_type(data_source(pkg = "qStates",
-                          database = "states",
-                          print = FALSE), "list")
-          })
-
-test_that("data_source() returns the correct output format at the dataset
-          level", {
-  expect_type(data_source(pkg = "qStates",
-                          database = "states",
-                          dataset = "GW",
-                          print = FALSE), "character")
-          })
-
-test_that("data_contrast() returns the correct output format at the package
-          level", {
-  expect_type(data_contrast(pkg = "qStates", print = FALSE), "list")
+test_that("data_source() returns the correct output format", {
+  expect_type(data_source("qData"), "list")
+  expect_type(data_source("qData", "emperors"), "list")
+  expect_type(data_source("qData", "emperors", "wikipedia"), "list")
 })
 
-test_that("data_contrast() returns the correct output format at the database
-          level", {
-  expect_type(data_contrast(pkg = "qStates",
-                            database = "states",
-                            print = FALSE), "list")
-})
-
-test_that("data_contrast() returns the correct output format at the dataset
-          level", {
-  expect_type(data_contrast(pkg = "qStates",
-                            database = "states",
-                            dataset = "GW",
-                            print = FALSE), "list")
+test_that("data_contrast() returns the correct output format", {
+  expect_type(data_contrast("qData"), "list")
+  expect_type(data_contrast("qData", "emperors"), "list")
+  expect_type(data_contrast("qData", "emperors", "wikipedia"), "list")
 })